chore: bump requests version and update geoclient baseurl to v2
This is a fork from an old library no longer actively maintained and has become outdated.
My hope is to polish this repo up and deploy it to AWS to automate address geocoding for the PEU Data Team.

Original git repo https://github.com/ishiland/python-geoclient

Update python requests library from v2.21.0 to v2.31.0 and updates the BASE_URL variable in the config file to v2.